Move provider into scope
Move 'provider' from the 'cloud/services.provider' package to 'scope'.
This is in preparation for adding ProviderScope.
